requirements.txt
psycopg2-binary
python-dotenv
telethon
pandas
ultralytics
Pillow
numpy
fastapi
uvicorn